# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/videos": {
            "post": {
                "description": "If the video is not yet tracked, the service:\n1) validates the URL/ID,\n2) fetches fresh stats from the provider,\n3) creates a new video record in the DB and writes the first stats snapshot.\nIf the video is already tracked, the service DOES NOT call the provider.\nIt returns the latest saved views and earnings from the `videos` table\nand also appends a new row to the hourly stats journal.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "TrackVideo TikTok video for tracking",
                "parameters": [
                    {
                        "description": "Video URL or TikTok ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TrackVideoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TrackVideoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid URL/ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Video not found on TikTok",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Provider rate limit",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/videos/{tiktok_id}": {
            "get": {
                "description": "Returns the last saved views and earnings for a TikTok video\nfrom the `videos` table. Does NOT call external provider.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Get latest saved TikTok video stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TikTok video ID",
                        "name": "tiktok_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TrackVideoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid TikTok ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Video not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.TrackVideoRequest": {
            "type": "object",
            "properties": {
                "tiktok_id": {
                    "type": "string",
                    "example": "1234567890"
                },
                "url": {
                    "type": "string",
                    "example": "https://www.tiktok.com/@user/video/1234567890"
                }
            }
        },
        "models.TrackVideoResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-11-24T01:30:00Z"
                },
                "currency": {
                    "type": "string",
                    "example": "USD"
                },
                "current_earnings": {
                    "type": "number",
                    "example": 1.5
                },
                "current_views": {
                    "type": "integer",
                    "example": 15000
                },
                "last_updated_at": {
                    "type": "string",
                    "example": "2025-11-24T01:30:00Z"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "tiktok_id": {
                    "type": "string",
                    "example": "1234567890"
                },
                "title": {
                    "type": "string",
                    "example": "My viral video"
                },
                "url": {
                    "type": "string",
                    "example": "https://www.tiktok.com/@user/video/1234567890"
                },
                "video_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}